+++ src/fsd_widgets/board/ui/BoardCreationForm/index.ts
@@ -1,0 +1,1 @@
+export { BoardCreationForm } from './BoardCreationForm';
