+++ src/widget/index.ts
@@ -4,7 +4,6 @@ export { SideBar } from './SideBar';
 export { CircleMemberManagementDetail } from './CircleMemberManagementDetail';
 
 // Board
-export { BoardCreationForm } from './board/BoardCreationForm';
 export { BoardHeader } from './board/BoardHeader';
 export { BoardPostList } from './board/BoardPostList';
